+++ src/com/apmath/loans/infrastructure/models/loans/LoanInitialization.kt
@@ -6,6 +6,6 @@ import com.apmath.loans.domain.models.loans.LoanInitializationInterface
 data class LoanInitialization(
     override val amount: Money,
     override val term: Int,
-    override val interest: Int,
+    override val interest: Float,
     override val date: String
 ) : LoanInitializationInterface
